backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9
Pillow==10.4.0
numpy==2.1.1
qdrant-client==1.12.1
pytest==8.3.3
httpx==0.27.2
